wip: moved native bench to an example file and tested with `flamegraph`
[ci skip]
