+++ Cargo.toml
@@ -22,3 +22,7 @@ ring = "0.16"
 
 [build-dependencies]
 napi-build = "2"
+
+# FIXME: included for flamegraph, remove later?
+[profile.release]
+debug = true
