+++ src/native/napi/lib.rs
@@ -8,8 +8,6 @@ mod connection;
 mod stream;
 mod path;
 mod packet;
-// FIXME: remote later, just here to include rust tests
-mod test;
 
 #[napi]
 pub fn version_is_supported(version: u32) -> bool {
